+++ PDF-Suite/src/pdfsuite/forms.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 PT-PT: Escrita dos campos no PDF.