+++ PDF-Suite/tests/test_money.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 PT-PT: Testes da interpretacao de numeros, moeda e IVA.